backend/Automatic-Tags/GoogleCloudFunctionNLP/requirements.txt
google-cloud-language==2.7.0
flask==2.0.1
